requirements.txt
cryptography
PyMySQL
python-dotenv
rich
pycryptodome
mysql-connector-python